Update DataManager.cs
Added new members: newPlayerName string to be used to store the player name recovered from the input field until the end of the round to pass it as parameter to the checkScore function.

PlayerNameIncidents enum, meant to be returned from the  setNewPlayerName function, but not used at the moment....

Edited SetNewPlayerName function body to check if the user has writen something or not. Just checks if there is somethin, no filtering. The max character count is set to 15 directly from the editor options of the InputField component

added Using UnityEngine.SceneManagement